refactor: share helpers
